# Add reservoir, fountain and well range when constructing