--- a/src/city/view.c
+++ b/src/city/view.c
@@ -557,8 +557,14 @@ static void do_valid_callback(int x, int y, int grid_offset, map_callback *callb
     }
 }
 
-void city_view_foreach_tile_in_range(int x, int y, int grid_offset, int size, int radius, map_callback *callback)
+void city_view_foreach_tile_in_range(int grid_offset, int size, int radius, map_callback *callback)
 {
+    int x, y;
+    city_view_grid_offset_to_xy_view(grid_offset, &x, &y);
+    x = (x - data.camera.tile.x) * TILE_WIDTH_PIXELS - (y & 1) * HALF_TILE_WIDTH_PIXELS - data.camera.pixel.x + data.viewport.x;
+    y = (y - data.camera.tile.y - 1) * HALF_TILE_HEIGHT_PIXELS - data.camera.pixel.y + data.viewport.y;
+    grid_offset += GRID_SIZE + 1;
+
     int x_delta = HALF_TILE_WIDTH_PIXELS;
     int y_delta = HALF_TILE_HEIGHT_PIXELS;
     int x_offset = HALF_TILE_WIDTH_PIXELS;
@@ -578,7 +584,7 @@ void city_view_foreach_tile_in_range(int x, int y, int grid_offset, int size, in
         do_valid_callback(x, y - y_offset, offset_north, callback);
         do_valid_callback(x - x_offset - x_delta, y, grid_offset + map_grid_delta(-ring - 2, ring + size), callback);
         do_valid_callback(x + x_offset + x_delta, y, grid_offset + map_grid_delta(ring + size, -ring - 2), callback);
-        for (int tile = 1; tile < (ring + size) * 2; ++tile) {
+        for (int tile = 1; tile < ring * 2 + size + 2; ++tile) {
             do_valid_callback(x + x_delta * tile, y - y_offset + y_delta * tile, offset_north + tile, callback);
             do_valid_callback(x - x_delta * tile, y - y_offset + y_delta * tile, offset_north + tile * GRID_SIZE, callback);
             do_valid_callback(x + x_delta * tile, y + y_offset - y_delta * tile, offset_south - tile * GRID_SIZE, callback);