--- a/src/city/view.h
+++ b/src/city/view.h
@@ -71,7 +71,7 @@ void city_view_foreach_map_tile(map_callback *callback);
 
 void city_view_foreach_valid_map_tile(map_callback *callback1, map_callback *callback2, map_callback *callback3);
 
-void city_view_foreach_tile_in_range(int x, int y, int grid_offset, int size, int radius, map_callback *callback);
+void city_view_foreach_tile_in_range(int grid_offset, int size, int radius, map_callback *callback);
 
 void city_view_foreach_minimap_tile(int x_offset, int y_offset, int absolute_x, int absolute_y, int width_tiles, int height_tiles, map_callback *callback);