… src/core/config.h
… src/window/config.c