Example:

![city 2020-03-02 08 32 34](https://user-images.githubusercontent.com/6518876/75658747-bdd9ba00-5c60-11ea-973d-d6d9adeee2bd.png)

Fix #343.

It shows the range whether the reservoir is next to water or not, and whether the fountain has reservoir access or not. Desert climate was also taken into account.

Also slightly improves texture alpha blending performance.