@crudelios 2020-03-02
Before doing that, I'm trying to improve the performance. It should be slightly faster, but no miracle yet...