@crudelios 2020-03-02
Fair enough. I'll do both.