@CommonLoon102 2020-03-02
> 
> 
> Even without the performance hit I'd rather have a setting for it.

The question is, should it be enabled or disabled by default?