@crudelios 2020-03-02
Can you test the performace of the new code? I also think I found a slight optimization to all `draw_compressed` routines.

> should it be enabled or disabled by default?

Since we're adding a setting that, when enabled, makes the game look different, I'd say it should be disabled by default.