@bvschaik 2020-03-03
> The question is, should it be enabled or disabled by default?

The default settings should look and act like plain Caesar 3, so disabled by default.

I'll have a look at the performance later today when I get back home.