@crudelios 2020-03-03
Really? I thought avoiding the constant calls to `graphics_get_pixel` would make things faster.

My visual studio profiler always crashes so I have to go by feeling :/