@theradeonxt 2020-03-03
Isn't it because some math done in that MIX   step? Also the way it is done per each byte, it's directly affevted by image size.
Even the get pixel function can be optimmized by compiler.