@crudelios 2020-03-03
Even though my profiler doesn't work, I tested this manually by using the windows-specific `QueryPerformanceCounter` and measuring the time between the start and the end of the function `city_view_foreach_tile_in_radius`.

I tested both the new mixing function and the array access independently.

Tests were performed using `mingw-w64` with `-O3` optimization. The same spot was used to display the building ghost.

For simplicity, I tested both the new and the old mixing functions using the original `dst` pointer dereference from the `master` branch.

Here are the times (averaged from a 10 second run each):

* Old alpha mixing function: 2.04ms
* New alpha mixing function: 1.56ms

Next, I changed the `dst` access when using the new alpha mixing function:

* Old style (from `master`): 1.56ms
* Array indexing (from most recent commit): 1.52ms
* New style: one call to `graphics_get_pixels` per line, `dst` incremented: 1.46ms

Curiously enough, I get a performance increase using the new code. However there is indeed more to gain by ditching the array indexing and using pointer incrementation along with a single `graphics_get_pixel` per row.

I'll also test at home where I have more available compilers for Windows.