@crudelios 2020-03-03
This latest implementation is the fastest one for me across all compilers. Can you test, please?