@bvschaik 2020-03-04
> This latest implementation is the fastest one for me across all compilers. Can you test, please?

Tested it, looks good 👍  Just one comment about the color.