@crudelios 2020-03-04
Out of curiosity I tried changing the number/order of `graphics_get_pixel` calls for `draw_compressed` to be similar to this implementation and the performance was marginally worse. So I don't think I'll be trying anything else regarding performance improvements.