@bvschaik 2020-03-04
I just did some testing with using premultiplied alpha, but performance also got worse, so I guess this is as good as we'll get using the current CPU-based setup. Thanks for profiling!